Use arrow-jdbc to extra the scott emp data set into ARROW format for testing
Write a few more tests for show where the execution occurs for the query adapter.

Thst vast majority of work is still being done at the enumerable layers, but it does appear
that certain calculations / projects can likely be pushed down further.

The current limitation with projection is that it only projects simple column refernces rather than
more complex calculations which is where much of the power still lies in gandiva.